Add geolocation and geocoding functionality
Add geolocation functionality in StoreAndRetrieveLocations which
stores the users location in a Firebase realtime database. This location
is also used in MapViewScreen where a marker is shown for the users
location

Add reverse geocoding functionality which takes the users geolocation
and turns it into an address using latitude and longitude coordinates.
The city and state is then stored in the Firebase realtime database

Add firebase querying functionality which uses the users city and state
to filter all users who are also inside of that city and these users are
returned

Minor fix which passes as a prop both the user's uid and username to
MapViewScreen from the auth functionality. This makes the transition
from Auth to Home more smooth and keeps MapViewScreen from having to
make unnecessary function calls to retrieve the username every time
a navigation is made back to MapViewScreen